Remove a (UINT) cast to silence a VC6 compiler warning (this was added by #27352, but Digest-SHA-5.37 changed the loop variable to be signed so the cast now needs to go). Also bump the VERSION since #27482 has not yet been applied to the CPAN version.
p4raw-id: //depot/perl@28630

+++ ext/Digest/SHA/src/sha.c
@@ -527,7 +527,7 @@ SHA *s;
 		return(0);
 	SHA_fprintf(f, "alg:%d\nH", s->alg);
 	for (i = 0; i < 8; i++)
-		for (j = 0; j < (UINT) (s->alg <= 256 ? 4 : 8); j++)
+		for (j = 0; j < (s->alg <= 256 ? 4 : 8); j++)
 			SHA_fprintf(f, "%s%02x", j==0 ? ":" : "", *p++);
 	SHA_fprintf(f, "\nblock");
 	for (i = 0; i < (int) (s->blocksize >> 3); i++)
